279-subscription-general: PR comments.
