+++ packages/orchestrator-ui-components/src/components/WFOSubscription/index.ts
@@ -1,8 +1,8 @@
 export * from './utils';
 export * from './WFOProcessesTimeline';
-export * from './SubscriptionProductBlock';
-export * from './SubscriptionActions';
+export * from './WFOSubscriptionProductBlock';
+export * from './WFOSubscriptionActions';
 export * from './SubscriptionKeyValueBlock';
-export * from './SubscriptionDetailTree';
-export * from './SubscriptionGeneral';
+export * from './WFOSubscriptionDetailTree';
+export * from './WFOSubscriptionGeneral';
 export * from './WFOSubscription';
