+++ packages/orchestrator-ui-components/src/components/WFOSubscriptionsList/subscriptionListTabs.ts
@@ -1,5 +1,6 @@
 import { WFOFilterTab } from '../../components';
-import { SubscriptionListItem } from '../../types';
+import { SubscriptionListItem } from './mapGrapghQlSubscriptionsResultToSubscriptionListItems';
+
 export enum WFOSubscriptionsTabType {
     ACTIVE = 'ACTIVE',
     TERMINATED = 'TERMINATED',
