+++ packages/orchestrator-ui-components/src/utils/getDefaultTableConfig.ts
@@ -15,8 +15,9 @@ import {
     ProductBlockDefinition,
     ProductDefinition,
     WorkflowDefinition,
-    SubscriptionListItem,
 } from '../types';
+
+import { SubscriptionListItem } from '../components/WFOSubscriptionsList';
 
 import { ProcessListItem } from '../components/WFOProcessesList/WFOProcessList';
